refactor: `typed-object` 폴더 생성 및 유틸 함수 이동
[#16]
