+++ src/utils/typed-object/getTypedObjectEntries.ts
