+++ src/utils/typed-object/getTypedObjectFromEntries.ts
